Decorating Rails helpers
In the source it seems to indicate that helpers aren't decoratable, due
to the weird way that Rails loads them. The `Decorator` source indicates
that there indeed is a programmatic way to do this, but unfortunately
this results in about 30 extra lines of configuration added when the app
boots. Additionally, the pattern of defining a module and then calling a
class method to mix it in with whatever object we want to manipulate is
pretty much how decorators work in the first place.

This removes the raised error when a `Helper` is decorated, and instead
utilizes the `ActionController::Base.helper`'s alternative syntax
wherein you can pass it a block and *that* is included as the helper.
We just pass the `mod` that is generated from `decorate` into the
`helper`, mimicing the way developers actually override helpers in the
real world.