Decorate mixin modules
Previously, mixins (defined as modules that get included somewhere)
could not be decorated because of complex load order scenarios that we
didn't want to tackle in a "v1" of this gem. However, this has always
been a bit of a wart in codebases trying to decorate mixed in code, and
for some reason can't keep track of the classes it's being mixed into at
any point in time. We accomplish this by introducing a
`Module#mixed_into` method which keeps track of classes that a module
gets included into. When a mixin module is decorated, the classes that
module is mixed into get decorated with the same module, in the same
order. This decoration also occurs whenever the module is included into
classes after the fact.